Store hue and saturation in the units the UI uses
Setting 195/90/70 and rebooting gave 194/89/70 back. Not a rounding error in
RGB - RGB is not involved. Settings kept hue and saturation scaled to a byte,
so every value crossed two truncating integer divisions, one on the way in and
one on the way out. Both round down, so the errors added up instead of
cancelling: 358 of 360 hues came back wrong, by up to 2, and 95 of 101
saturations by 1.

The two halves of that were different problems. 360 hues genuinely do not fit
in 256 steps. Saturation never had that excuse - 101 values into 256 slots is
upsampling, and it was losing them purely to truncation.

Both are stored in the UI's own units now (uint16_t 0..359, byte 0..100), which
makes the round trip exact for every one of the 360 x 101 combinations. The
conversion to the 8 bits FastLED wants happens once, where the colour reaches
the driver, and rounds to the nearest step. Nothing changes at the LEDs: they
only ever had 256 hue steps, and 195/90 still lands on CHSV(139, 230).

Existing records are converted rather than misread. fillDocument stamps
SETTINGS_SCHEMA, and a record without that field is schema 1 and gets scaled up
on load - the stored 138/229 comes back as 194/90, which is what that clock was
actually showing. WiFi credentials are untouched by any of this: they live in
the WiFi stack's own NVS namespace, not in ours.

+++ src/Settings.h
@@ -40,10 +40,13 @@ public:
     boolean getUseLdr() {return this->UseLdr;}
     void    setBrightness(byte Brightness) {this->Brightness = Brightness;}
     byte    getBrightness() {return this->Brightness;}
-    void    setColorHue(byte ColorHue) {this->ColorHue = ColorHue;}
-    byte    getColorHue() {return this->ColorHue;}
-    void    setColorSat(byte ColorSat) {this->ColorSat = ColorSat;}
-    byte    getColorSat() {return this->ColorSat;}
+    // Hue 0..359 and saturation 0..100 - the units the web UI works in, so a
+    // value set there comes back unchanged. The conversion to the 8 bits
+    // FastLED wants happens once, where the colour is handed to the driver.
+    void     setColorHue(uint16_t ColorHue) {this->ColorHue = ColorHue;}
+    uint16_t getColorHue() {return this->ColorHue;}
+    void     setColorSat(byte ColorSat) {this->ColorSat = ColorSat;}
+    byte     getColorSat() {return this->ColorSat;}
     void    setMode (byte Mode) {this->Mode = Mode;}
     byte    getMode() {return this->Mode;}
     char*   getNTPServer() {return this->NTPServer;}
@@ -99,9 +102,9 @@ private:
     boolean RenderCornersCw;
     boolean RenderColorCorner;
     boolean UseLdr;
-    byte    Brightness;
-    byte    ColorHue;
-    byte    ColorSat;
+    byte     Brightness;   // 0..100
+    uint16_t ColorHue;     // 0..359
+    byte     ColorSat;     // 0..100
     byte    Mode;
     
     // Entries for automatic Timezone
